removed interfaces and renamed base classes to interface names; implemented ArtifactGet using prepared statement to avoid SQL injection via unsafe content (Artifact.uri)
